Move display_ops.py to _internal
